+++ packages/admin-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies [1147b22]
+  - @shopware-ag/meteor-component-library@4.12.1
+
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies [215eec6]
+  - @shopware-ag/meteor-component-library@4.12.0
+
 ## 6.0.0
 
 ### Major Changes